% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_538.R
\name{style_538}
\alias{style_538}
\title{Add FiveThirtyEight theme to ggplot chart}
\usage{
style_538()
}
\description{
This function allows you to add the FiveThirtyEight theme to your ggplotgraphics.
}
\examples{
library(ggplot2)
hist <- ggplot(diamonds) +
geom_histogram(aes(x = price), fill = "#ed713a") +
style_538()
}
\keyword{style_538}
